% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allocate_news}
\alias{allocate_news}
\title{Function to attribute news items to the correct trading day}
\usage{
allocate_news(d_n, nyse = TRUE, eod_h = 16, eod_tz = "EST", ind = NULL)
}
\description{
Function to attribute news items to the correct trading day
}
