% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_hist_members}
\alias{clean_hist_members}
\title{NOT FOR EXPORT. Idiosyncratic function to tidy index membership xts.}
\usage{
clean_hist_members(hist_members)
}
\arguments{
\item{hist_members}{xts object containing membership data. Column names are
assumed to be asset names corresponding to those in df. Note: All entries
not NA are assumed to represent active membership.}
}
\description{
Idiosyncratic function to clean symbol names of index
membership xts and replace names which had changed between data gathering
 and index membership update. Only created to maintain reproduceability.
}
