% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_names}
\alias{clean_names}
\title{Function to clean individual asset OHLC data.}
\usage{
clean_names(symbolList, totalReturn = FALSE, Member = FALSE, ind = NULL)
}
\arguments{
\item{symbolList}{List of symbols loaded into the environment.}

\item{totalReturn}{If TRUE, will look for column of form "total_return", and
will use for adjustment.}

\item{Member}{If TRUE, will preserve membership column.}
}
\description{
Function to clean individual asset OHLC data. Changes column
names, and provides option to use adjusted data column to weight OHLC prices.
Silently drops all columns except for OHLC and Member (if Member=TRUE).
}
