% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daily_news_scores}
\alias{daily_news_scores}
\title{Function to calculate daily news scores from xts of news data}
\usage{
daily_news_scores(dn, min_relevance = 0.6, news_type = "all", ind = NULL)
}
\description{
Function to calculate daily news scores from xts of news data
}
