% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols}
\alias{get_cols}
\title{Function for extracing a given column from loaded xts objects and combining
them into a single object.}
\usage{
get_cols(symbolList, cname)
}
\description{
Function for extracing a given column from loaded xts objects and combining
them into a single object.
}
