% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_firm_data}
\alias{get_firm_data}
\title{Function to extract firm variables for a list of assets over a specified
formation period and save in individual .csv files.}
\usage{
get_firm_data(firm_saveDir, firmDir, priceDir, assets, ind = NULL)
}
\description{
Function to extract firm variables for a list of assets over a specified
formation period and save in individual .csv files.
}
\note{
Not for export.
}
