% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_news_vars}
\alias{get_news_vars}
\title{Function to extract news variables for a list of assets over a specified
formation period and save in individual .csv files.}
\usage{
get_news_vars(newsDir, saveDir, assets, n = 1, ind, eod_h = 16,
  eod_tz = "EST", on = "months")
}
\description{
Function to extract news variables for a list of assets over a specified
formation period and save in individual .csv files.
}
\note{
Not for export.
}
