% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{get_weights_ls}
\alias{get_weights_ls}
\title{Function to get equal-weighted long-short position/weight vector from ranks.}
\usage{
get_weights_ls(ranks, TopN)
}
\arguments{
\item{ranks}{xts object containing asset ranks. Column names are assumed to
be independent assets.}

\item{TopN}{Number of assets on each side (long or short) of trade (integer).
Total number of assets held each period will be TopN*2
#' @importFrom zoo index}
}
\description{
Takes an xts of ranks and returns an xts object containing
position sizes (weights) for an equal weighted long-short portfolio.
}
