% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{holdingTime}
\alias{holdingTime}
\title{Function returns logical if holding time exceeds the simulation end date.}
\usage{
holdingTime(current_date, end_date, holding_time, holding_unit = "month")
}
\description{
Function returns logical if holding time exceeds the simulation end date.
}
