% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_files}
\alias{load_files}
\title{Function to read csv files and load them as xts into selected environment.}
\usage{
load_files(dir, fnames, index.col = "date", daterange = NULL,
  envir = globalenv(), delim = ",")
}
\arguments{
\item{dir}{The directory in which the required files are stored (string)}

\item{index.col}{Column name to be referenced for ordering xts object
(string). Must be coercible to POSIXct.}

\item{daterange}{The date range of resulting xts to be loaded (string).}

\item{envir}{The environment in which to load the objects.}

\item{fname}{name of file to be loaded (string)}
}
\description{
Function to load multiple csv files into the selected environment as xts
objects. Object names are the file names, minus the '.csv' extension.
}
