% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{momRankMembers}
\alias{momRankMembers}
\title{Function for ranking assets on momentum that require index membership for
consideration.}
\usage{
momRankMembers(df, n = 6, s = 0, hist_members = newHM, on = "months",
  ind = NULL)
}
\arguments{
\item{df}{xts object containing price series of assets to rank. Every column
will be treated as an independent asset. Column names of assets to rank must
also appear in column names of `hist_members`}

\item{n}{Formation period e.g. '6' with `on`=months' for 6 month formation
period. (integer)}

\item{s}{Skip period (integer). This is the time between the end of formation
period and ranking (and presumably investment) date).}

\item{hist_members}{xts object containing membership data. Column names are
assumed to be asset names corresponding to those in df. Note: All entries
not NA are assumed to represent active membership.}

\item{on}{Desired return periodicity (string). Passed to xts::endpoints.}
}
\description{
Function to rank assets based on price momentum, but only
consider those that have index membership at *time of ranking*. Function
performs ranking on last *trading day* of period. Future versions to include
the option to perform rankings at different days of period/month.
}
\note{
NEED TO REMOVE Automatic NYSE trading days.
}
