% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MonthlyFirmData}
\alias{MonthlyFirmData}
\title{Construct end-of-month measures of firm data.}
\usage{
MonthlyFirmData(f_d, d_p, ind = NULL)
}
\description{
Construct end-of-month measures of firm data.
}
\note{
Not for export!
}
