% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{monthlyNews_EOD}
\alias{monthlyNews_EOD}
\title{Dunction to creates news variables over n-month horizons. News items
released after COB are counted in next period.}
\usage{
monthlyNews_EOD(master_ind, d_n, eod_h = 16, eod_tz = "EST", n = 1,
  on = "months")
}
\arguments{
\item{master_ind}{Date or POSIXct array that will be used to determine
xta endpoints. Should reflect trading days.}

\item{d_n}{xts of TRNA data (single entity).}

\item{eod_h}{Local time (24 hr format) of exchange close-of-business (int).}

\item{eod_tz}{Time zone in which eod_h is specified (string).}

\item{n}{Formation period (months).}

\item{on}{Desired return periodicity (string). Passed to xts::endpoints.}
}
\description{
Dunction to creates news variables over n-month horizons. News items
released after COB are counted in next period.
}
\note{
Not for export!
}
