% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naFillFrom}
\alias{naFillFrom}
\title{Function to replace NAs on last trading day of month with
value from previous trading day.}
\usage{
naFillFrom(df, firstDay = FALSE, fromNext = FALSE, ndays = 2,
  ind = NULL, on = "months")
}
\arguments{
\item{df}{xts object containing price series of assets to rank. Every column
will be treated as an independent asset.}

\item{firstDay}{If TRUE, replaces NAs on first trading day of month.}

\item{fromNext}{If True, replaces NAs with value from next business day.}

\item{ndays}{The maximum number of days from NA from which to find data.}
}
\description{
Function to replace NAs on last trading day of month with
value from previous or next trading day, up to `ndays`. Only replaces
data on endpoints (option for start points). Within-period NAs will not be
filled.
}
\note{
NEED TO REMOVE Automatic NYSE trading days.
}
