% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{news_agg}
\alias{news_agg}
\title{Make to handle combined news xts?? That way can use centered vars
Currently weights each sentiment day for each asset by relative the recency
of the sentiment day with respect to the assets own news history *not*
relative to absolute recency.}
\usage{
news_agg(dn, n = 1, w = 1, on = "months", ind = NULL,
  from_first = TRUE)
}
\arguments{
\item{dn}{combined xts of daily news measure for multiple assets. e.g.
get_cols applied to output of `daily_news_scores`}
}
\description{
Make to handle combined news xts?? That way can use centered vars
Currently weights each sentiment day for each asset by relative the recency
of the sentiment day with respect to the assets own news history *not*
relative to absolute recency.
}
