% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nyse_trading_dates}
\alias{nyse_trading_dates}
\title{Function to generate index of all NYSE trading dates.}
\usage{
nyse_trading_dates(start_date, end_date, freq = "days", tz = "UTC")
}
\description{
Function to generate index of all NYSE trading dates.
}
