% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{place_transactions}
\alias{place_transactions}
\title{Blotter portfolio simulation function that does not consider index
membership. Analogous to `place_transactions_index`.}
\usage{
place_transactions(initEq, assets, Currency, ind, start_i = 1, K_m = 1,
  wts = wts, initdate = initdate, enddate = enddate,
  enter_prefer = "Close", exit_prefer = "Close", holding_time = FALSE)
}
\description{
Blotter portfolio simulation function that does not consider index
membership. Analogous to `place_transactions_index`.
}
