% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{place_transactions_index}
\alias{place_transactions_index}
\title{Function to simulate an individual (1 of K) momentum portfolio.}
\usage{
place_transactions_index(initEq, assets, Currency, ind, start_i = 1,
  K_m = 1, wts = wts, initdate = initdate, enddate = enddate,
  enter_prefer = "Close", exit_prefer = "Close")
}
\description{
Function to simulate an individual (1 of K) momentum portfolio.
This function assumes rebalancing/exits occur on same day of entries,
and does not look for exit explict exit/rebalnce signals. The exception is
for memberships.
}
\details{
This procedure assumes transactions all occur on same date, with
last available price (otherwise, the last trading date or memberhsip exit
date must be known in advance, or sorted within the procedure, since blotter
requires transactions to be added in date order).
}
\note{
Need to update now that `weights_i` function handles holdingTime.
}
