% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{place_transactions_ordering}
\alias{place_transactions_ordering}
\title{Similar to `place_transactions_index` but transacts on the last
traded price/membership date.}
\usage{
place_transactions_ordering(initEq, assets, Currency, ind, start_i = 1,
  K_m = 1, wts = wts, initdate = initdate, enddate = enddate,
  txn_data = NULL, enter_pefer = "Close", exit_prefer = "Close")
}
\description{
Similar to `place_transactions_index` but transacts on the last
traded price/membership date.
}
