% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{portfolio_cumulative.return}
\alias{portfolio_cumulative.return}
\title{Function to calculate portfolio equity given weights. Unlike the version
from PerformanceAnalytics, can handle long-short portfolios. Returns equity
curve as opposed to returns.}
\usage{
portfolio_cumulative.return(df, weights = NULL, rebalance_on = NULL,
  verbose = FALSE, zero_cut = TRUE)
}
\description{
Function to calculate portfolio equity given weights. Unlike the version
from PerformanceAnalytics, can handle long-short portfolios. Returns equity
curve as opposed to returns.
}
