% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readxts2}
\alias{readxts2}
\title{Function to read csv files as xts objects.}
\usage{
readxts2(dir, fname, index.col = "Date", daterange = NULL, delim = ",")
}
\arguments{
\item{dir}{The directory in which the required file is stored (string)}

\item{fname}{name of file to be loaded (string)}

\item{index.col}{Column name to be referenced for ordering xts object
(string). Must be coercible to POSIXct.}

\item{daterange}{The date range of resulting xts to be loaded (string).}
}
\value{
An xts object.
}
\description{
Function to read csv files as xts objects.
}
