% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_parallel}
\alias{run_parallel}
\title{Function to perform commands in parallel}
\usage{
run_parallel(n, ptext, chunk_size = NULL)
}
\description{
Function to perform commands in parallel
}
\examples{
DF_cl <- get_cols(sp500,"Close")
ptext <- "naFillFrom(df=DF_cl[,i_1:i_2], ind=master_ind)"
DF_cl <- run_parallel(n=ncol(DF_cl), ptext=ptext)
}
