% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{txn_dates2}
\alias{txn_dates2}
\title{Need to double check!!}
\usage{
txn_dates2(df, membership = TRUE, hist_members = newHM, on = "months",
  ind = NULL)
}
\description{
Need to double check!!
}
\note{
Much of the apparant complexity/ugliness in this fuction is to deal
with the case in which there is no price data in the current month. We could
just return an NA in all these cases, but there is the special case in which
this is not ideal: If the last day with membership and/or price data falls on
the end point of the previous period. Because positions are often evaluated
and entered based on end-of-period points, there is the possibility that we
could have entered such a position, and then have no reference point for
exit price or date the next period when we realise there is no longer
 membership or data.
}
