% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txn_pretty}
\alias{txn_pretty}
\title{Function to extract and neatly present the transactions of a list of
portfolios loaded into the environment.}
\usage{
txn_pretty(portfolio_list)
}
\description{
Function to extract and neatly present the transactions of a list of
portfolios loaded into the environment.
}
