% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{visualise_trades}
\alias{visualise_trades}
\title{Function for visualising trades of overlapping J/S/K portfolios
Mainl created for small number of entry/exit assets.}
\usage{
visualise_trades(weights, K_m, exits = TRUE)
}
\description{
Function for visualising trades of overlapping J/S/K portfolios
Mainl created for small number of entry/exit assets.
}
