% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_prep.R
\name{weights_i}
\alias{weights_i}
\title{Given an xts of portfolio weights with signals each period,
returns the kth portfolio weights.}
\usage{
weights_i(wts, start_i, k, holding_time = FALSE)
}
\arguments{
\item{holding_time}{If TRUE, entry weights only set if there is enough time
left to fulfill entire holding period.}
}
\description{
Given an xts of portfolio weights with signals each period,
returns the kth portfolio weights.
}
\note{
might be worth including enddate as an argument rather than using the
last date in mom. This aligns with the `place_transaction` functions.
}
